split mp_log_u32 for more configurability
